chore: mv getAggregatedRecords to separate tags action
